add spanish interface

+++ app/actions/settingActions.js
@@ -3,54 +3,57 @@ import {
     setTheme,
     setSendCTRLandEnter,
     setClock,
-    setDefaultSettings
+    setDefaultSettings,
+    setInterfaceLanguage,
+    setLanguage
 } from './actionCreators'
 import {
     USERNAME,
     THEME,
     CLOCK,
-    SEND_WITH_CTRL_AND_ENTER
+    SEND_WITH_CTRL_AND_ENTER,
+    LANGUAGE_TAG,
+    defaultLanguageValue
 } from "../constant/availableSettings"
-import {  themeLight } from "../constant/availableSettings"
-import {notifyConnection} from './messageActions'
+import { themeLight } from "../constant/availableSettings"
+import { notifyConnection } from './messageActions'
+import { readFromLocalStorage, writeOnLocalStorage } from '../utils/storageUtils'
+import { loadLanguage } from './languageActions'
 
 const handleSettingsChange = (event) => {
     return (dispatch) => {
         switch (event.target.name) {
             case USERNAME:
                 dispatch(setUsername(event.target.value))
-                localStorage.setItem(USERNAME, event.target.value)
+                writeOnLocalStorage(USERNAME, event.target.value)
                 break;
             case THEME:
                 dispatch(setTheme(event.target.value))
-                localStorage.setItem(THEME, event.target.value)
+                writeOnLocalStorage(THEME, event.target.value)
                 break;
             case CLOCK:
                 dispatch(setClock(parseInt(event.target.value)))
-                localStorage.setItem(CLOCK, event.target.value)
+                writeOnLocalStorage(CLOCK, event.target.value)
                 break;
             case SEND_WITH_CTRL_AND_ENTER:
                 dispatch(setSendCTRLandEnter(event.target.value == "true"))
-                localStorage.setItem(SEND_WITH_CTRL_AND_ENTER, event.target.value == "true")
+                writeOnLocalStorage(SEND_WITH_CTRL_AND_ENTER, event.target.value == "true")
+                break;
+            case LANGUAGE_TAG:
+                dispatch(loadLanguage(event.target.value))
+                writeOnLocalStorage(LANGUAGE_TAG, event.target.value)
                 break;
         }
     }
 }
 
 const loadSettings = () => {
-    if (!localStorage[USERNAME])
-        localStorage.setItem(USERNAME, "anonymous" + new Date().getMilliseconds())
-    if (!localStorage[THEME])
-        localStorage.setItem(THEME, themeLight)
-    if (!localStorage[CLOCK])
-        localStorage.setItem(CLOCK, 24)
-    if (!localStorage[SEND_WITH_CTRL_AND_ENTER])
-        localStorage.setItem(SEND_WITH_CTRL_AND_ENTER, true)
     return {
-        username: localStorage[USERNAME],
-        theme: localStorage[THEME],
-        clock: localStorage[CLOCK],
-        sendCTRLandEnter: localStorage[SEND_WITH_CTRL_AND_ENTER]
+        username: readFromLocalStorage(USERNAME, "anonymous" + new Date().getMilliseconds()),
+        theme: readFromLocalStorage(THEME, themeLight),
+        clock: readFromLocalStorage(CLOCK, 24),
+        sendCTRLandEnter: readFromLocalStorage(SEND_WITH_CTRL_AND_ENTER, true),
+        language: readFromLocalStorage(LANGUAGE_TAG, defaultLanguageValue)
     }
 }
 
@@ -63,13 +66,17 @@ const initializeSettings = () => {
         dispatch(setTheme(settings.theme))
         dispatch(setClock(parseInt(settings.clock)))
         dispatch(setSendCTRLandEnter(settings.sendCTRLandEnter == "true"))
+        dispatch(setLanguage(settings.language))
+        dispatch(loadLanguage(settings.language))
     }
 }
 
 const clearSettings = () => {
-    localStorage.clear()
+    if (localStorage)
+        localStorage.clear()
     return (dispatch) => {
         dispatch(setDefaultSettings())
+        dispatch(loadLanguage(defaultLanguageValue))
     }
 }
